+++ assistants/config/__init__.py
@@ -48,7 +48,7 @@ def get_config() -> Config:
         OPENAI_API_KEY=os.environ.get(custom_keys.ASSISTANTS_API_KEY_NAME, None),
         ANTHROPIC_API_KEY=os.environ.get(custom_keys.ANTHROPIC_API_KEY_NAME, None),
         DEFAULT_MODEL=os.environ.get("DEFAULT_MODEL", "gpt-4o-mini"),
-        CODE_MODEL=os.environ.get("CODE_MODEL", "o1-mini"),
+        CODE_MODEL=os.environ.get("CODE_MODEL", "o3-mini"),
         IMAGE_MODEL=os.environ.get("IMAGE_MODEL", "dall-e-3"),
         ASSISTANT_INSTRUCTIONS=os.environ.get(
             "ASSISTANT_INSTRUCTIONS", "You are a helpful assistant."